New math library from scratch that enables the use of the YMM registers on AVX2-supported CPUs. Application in a runnable state, but MxLookAt and MxLookTo functions not implemented at the moment.

+++ project-assist-build-version-0.5-indev/Light.h
@@ -12,28 +12,28 @@ public:
 	struct LightData {
 		
 		struct {
-			float x{ 0.0f };
-			float y{ 0.0f };
-			float z{ 0.0f };
-			float w{ 1.0f };
+			double x{ 0.0f };
+			double y{ 0.0f };
+			double z{ 0.0f };
+			double w{ 1.0f };
 		} pos;
 		struct {
-			float x{ 1.0f };
-			float y{ 1.0f };
-			float z{ 1.0f };
-			float w{ 1.0f };
+			double x{ 1.0f };
+			double y{ 1.0f };
+			double z{ 1.0f };
+			double w{ 1.0f };
 		} direction;
 		struct {
-			float r{ 1.0f };
-			float g{ 1.0f };
-			float b{ 1.0f };
-			float a{ 1.0f };
+			double r{ 1.0f };
+			double g{ 1.0f };
+			double b{ 1.0f };
+			double a{ 1.0f };
 		} color;
 
-		float spotAngle{ 15.0f };
-		float constAtten{ 1.0f };
-		float linAtten{ 0.08f };
-		float quadAtten{ 0.0f };
+		double spotAngle{ 15.0f };
+		double constAtten{ 1.0f };
+		double linAtten{ 0.08f };
+		double quadAtten{ 0.0f };
 
 		int type{ POINT_LIGHT };
 		int isEnabled{ false };
