+++ project-assist-build-version-0.5-indev/Timer.cpp
@@ -3,14 +3,14 @@
 Timer::Timer() {
 	previous = std::chrono::steady_clock::now();
 }
-float Timer::mark() {
+double Timer::mark() {
 	const auto before = previous;
 	previous = std::chrono::steady_clock::now();
-	const std::chrono::duration<float> frameDuration = previous - before;
+	const std::chrono::duration<double> frameDuration = previous - before;
 	return mSpeedMult * frameDuration.count();
 }
-float Timer::peek() const {
-	return mSpeedMult * std::chrono::duration<float>(std::chrono::steady_clock::now() - previous).count();
+double Timer::peek() const {
+	return mSpeedMult * std::chrono::duration<double>(std::chrono::steady_clock::now() - previous).count();
 }
 void Timer::incSpeedMult() noexcept {
 	mSpeedMult += 1.0f;
